+++ cypress/integration/WidgetsForEnrollmentPages/WidgetProfile/index.js
@@ -1,3 +1,4 @@
+import '../../sharedSteps';
 
 Then('the profile details should be displayed', () => {
     cy.get('[data-test="profile-widget"]')
@@ -20,3 +21,13 @@ Then('the widget profile should be closed', () => {
         });
 });
 
+Then(/^the user sees the edit profile modal/, () =>
+    cy.get('[data-test="modal-edit-profile"]').within(() => {
+        cy.contains('Edit Person').should('exist');
+        cy.contains(
+            'Change information about this Person here. To change information about this enrollment, use the Edit button in the in the Enrollment box on this dashboard',
+        ).should('exist');
+        cy.contains('Save changes').should('exist');
+        cy.contains('Cancel without saving').should('exist');
+    }),
+);