+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/FormFoundation/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { build as buildFormFoundation } from './RenderFoundation';