+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { DataEntryProfile as DataEntry } from './DataEntry.container';